backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "bcrypt>=4.0.0",
    "grpcio>=1.74.0",
    "grpcio-reflection>=1.74.0",
    "grpcio-tools>=1.74.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.43",
]
